requirements.txt
streamlit
openai
instructor
pandas
atomic-agents
fastapi
uvicorn
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
requests
jinja2
